+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Do not fail if version check pattern doesn't match; instead assume that the binary is not installed.
     This supports multiple variants of the same binary, e.g. Hugo and Hugo Extended (see [GH-10]).
+- Sort output of `list`, `installed` and `outdated` by name.
 
 [GH-10]: https://github.com/lunaryorn/homebins/issues/10