+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.5] – 2020-07-31
+
 ### Changed
 
 - Do not fail if version check pattern doesn't match; instead assume that the binary is not installed.
@@ -51,4 +53,5 @@ Initial prerelease.
 [0.0.2]: https://github.com/lunaryorn/homebins/compare/v0.0.1...v0.0.2
 [0.0.3]: https://github.com/lunaryorn/homebins/compare/v0.0.2...v0.0.3
 [0.0.4]: https://github.com/lunaryorn/homebins/compare/v0.0.3...v0.0.4
-[Unreleased]: https://github.com/lunaryorn/homebins/compare/v0.0.4...HEAD
+[0.0.5]: https://github.com/lunaryorn/homebins/compare/v0.0.4...v0.0.5
+[Unreleased]: https://github.com/lunaryorn/homebins/compare/v0.0.5...HEAD